requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aniso8601==7.0.0
appdirs==1.4.4
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
beautifulsoup4==4.12.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.3.0
cryptography==42.0.5
cycler==0.12.1
daphne==4.1.2
defusedxml==0.7.1
Django==3.2.18
django-apscheduler==0.6.2
django-channels-graphql-ws==0.9.1
django-cors-headers==4.3.1
django-crontab==0.7.1
django-filter==24.2
django-graphql-jwt==0.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
fonttools==4.54.1
frozendict==2.4.2
frozenlist==1.4.1
fuzzywuzzy==0.18.0
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==22.0.0
html5lib==1.1
httplib2==0.22.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
install==1.3.5
jaro-winkler==2.0.3
joblib==1.4.0
kiwisolver==1.4.7
lxml==5.2.1
matplotlib==3.9.2
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==24.0
pandas==1.5.3
pandas-datareader==0.10.0
peewee==3.17.2
pillow==10.3.0
promise==2.3
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.1
regex==2024.4.16
reportlab==3.6.11
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
Rx==1.6.3
schedule==1.2.1
scipy==1.14.1
service-identity==24.1.0
singledispatch==4.1.0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
text-unidecode==1.3
tqdm==4.66.2
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
whitenoise==6.6.0
xlrd==2.0.1
yarl==1.9.4
yfinance==0.2.38
zope.interface==6.3
